Stop right-padding cutters!
If we prefix them with a LOW_CHAR spacer, we don't actually need
to right-pad them. Which means we don't need to limit to a certain
cutter width, and we don't need to pass in a separate pad char
for bottomout.

But mainly, this is paving the way to try to handle cutter suffixes
in a reasonable way.

It also keeps the overall length probably smaller, and will help
us add support for 'date or other number' too.