Handle cutter suffix letters, woot.
Still pretty messy, should clean up code a bit.

Need to change cutter prefix from space to period, so we can
have a 'lower' intermediate separator for suffix, the hyphen.

None of this low-ascii punctuation is allowed by our regex
in the parsed portion of the cutter, so this should not cause
problems.